@@ -1,4 +1,4 @@
-import PersonalityQuiz from '@scripts/h5p-personality-quiz';
+import PersonalityQuiz from '@scripts/h5p-personality-quiz.js';
 
 // Load library
 H5P.PersonalityQuiz = PersonalityQuiz;
